Removing util class for the SparkMAX motor controllers, updating constants
SparkMAX motor controllers no longer have to pass through motors through a util class, this is due to them not being needed and can operate without being added to a child.

Made changes to add all the VictorSPX motor controllers to be using the CAN bus rather than the PWM control. Removed castings to send as they are not communicating over PWM.

Added more documentation to the Constants file for each class.

Changed Lift to Tilt as there was already a lift system named Winch, this was for clarity reasons.

Minor changes: 

- Made variables for the MechanumDriveTrain.drivetrain function clearer
- Added a unused VictorSPX motor controller to the Indexer, this was left out on accident.