Started work on Ball intake, clocking, and shooting systems, encoders
Ball intake, clocking, and shooter mechanism have started progress in how systems should be set up. Basic boilerplate code is present for now until fine-tuning and work can be done to optimize each system.

The MechanumDriveTrain now has proper classes for the encoders, it is not going directly back to the RoboRIO. It will be going to the SparkMAX motor controller and pulling all the data through that device. There are now methods to get position and velocity from the encoders, however these functions are not used as of yet. There will be future use for vitals and autonomous.

The shooter mechanism now has proper encoders as it was similar situation with the drivetrain. There are functions that are not used for shooting the ball out.

Added SmartDashboard data of all the important subsystems.

Removed all mentions of addChild due to not being used for anything important.

Fixed up encoder usages due to default parameters thinking its used as a Hall Effect sensor rather than what we are actually using which is Quadrature encoders. Put in boilerplate encoder values to be changed later. 

Working on getting analog sensors working, this will crash the simulation. Working on a workaround to get the simulation running.