Commands and subsystems for the elevator mechanisms are setup
Mechanisms for the elevator and commands for the robot to perform actions are now set up. There are two simple functions that at the moment are not called that will winch the robot up and down and tilt the robot "up" and "down".

Control schemes for the robot will be handled on the commands and will be most likely assisted with the help of microswitches to help stop the motors from burning up.