Commands and subsystems for the ball mechanisms are setup
Commands and subsystems for the ball intakes, shooter, and clocking mechanisms are all set up to use the hardware. Some new functions for the ball intake are created and are largely boilerplate.

Commands are not yet assigned to button controls and will be implemented later.

Linted files that were not already linted to adhere to coding standards.