Added the built in Accelerometer, added documentation and power constants
Added methods to pull the Accelerometer built-in to the RoboRIO. This will help with stability when the robot does climb with the Winch and Tilt subsystems.

Added documentation to the Clocking, Intake, Tilt, and Winch subsystems. Documentation is still not final so expect more changes.

Added more power constants. The power constants are not used universally however this will be cleaned up and finalized.

Updated the .gitignore to remove files related to Simulation. This was proved to be a pain to manage to they were purged from the Git repository.