Add some controls to the robot, adding some utilities from current source
Adding some form of controlling mechanisms to the robot, this includes some tilting and winching actions to be performed. The rest of the controls will be added soon once a proper diagram is made on how to control the robot.

Added a utility that sets the power of the motor to the correct output from calculating the velocity needed to make the ball at any distance from the field. The current codebase for this was pulled from the public repository and is currently under massive restructuring for documentation purposes and validation that this is remotely possible to begin with. This involves use of a Limelight sensor but I am currently unaware of this being put on the robot, much less actually being used.

Removed creating and calling individual controllers and created methods to get global controllers at specified ports. This cuts down on code complexity and readability.