Add Power Distribution subsystem, linted document
The power distribution panel is now an entire subsystem. This will be used to pull vitals and other means of robot status to ensure the robot is running normally. There are functions that pull certain bits of data as well as pulls other data from what channel is inputted into the functions.

Linted the whole document, and changed how arrays are returned, instead of just creating a local variable and returning that, it is now creating a new array and returning that without a named variable.