Add global binds, renaming subsystems, added docs
Added global button binds for specific functions such as shooting the ball and setting the ball to run at specified power. Right now this will be running with full speed however this will change with other changes that will be in development.

Added a function to release the ball into the shooter mechanism, this will set the shooter motor to run at full tilt however this will change with other things that will be in development.

Added some more documentation with controllers being used in the commands.