Fix encoder values, removed throttle, fixed drivetrain
Fixed encoder values to use the correct number of counts per revolution. This will be used to help with speed and distance.

Removed throttle controls, as for the drivetrain to use the throttle controls is not going to be widely used.

Fixed controls going into the robot, this would cause a crash as the controller would return null, this is now fixed. Fixed a potential issue with the robot also taking the y-axis negative.