Use power constants rather than hardcoded values for subsystems
PowerConstants have been updated to fix a typo with the PowerDistributionPanel CAN ID, as well as updating the values for the power percentage for the intake, clocking, shooting, tilt, and winching subsytems. This will standardize power values for a single file to edit most power values for all the subsystems.