Removing calls to camera hardware to temporarily mitigate runtime error
There is an issue with the code that handles video, there is an improper handle that is causing the code to crash on startup due to how its being called into the robot container.

Removing the calls to the camera hardware is the simple fix for now until there is more concrete work put into researching how to get a camera to operate properly.